+++ example/index.html
@@ -42,11 +42,6 @@
       </div>
       <div class="row">
         <div class="col-md-12 col-xs-12 col-sm-12 col-lg-12">
-          <a href="record.html">Record CRUD</a>
-        </div>
-      </div>
-      <div class="row">
-        <div class="col-md-12 col-xs-12 col-sm-12 col-lg-12">
           <a href="user.html">User discovery</a>
         </div>
       </div>
