--- example/push.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="UTF-8">
-  <title>Skygear-SDK-JS</title>
-</head>
-<body>
-  <div id="main">
-    <p>
-      <label>End Point</label>
-      <span id="endpoint"></span>
-    </p>
-    <p>
-      <label>Token</label>
-      <span id="accessToken">
-        <a href="index.html">Go to Login</a>
-      </span>
-    </p>
-  </div>
-  <hr>
-  <script src="/bundle.js"></script>
-  <script type="text/javascript">
-    var m = document.getElementById("endpoint");
-    var token = document.getElementById("accessToken");
-    skygear.config({
-      'endPoint': '{{ SKYGEAR_ENDPOINT }}',
-      'apiKey': '{{ SKYGEAR_API_KEY }}'
-    }).then(function(container) {
-      m.textContent = container.endPoint;
-      token.innerText = container.auth.accessToken;
-    }, function(err) {
-      console.log(err);
-    });
-  </script>
-</body>
-</html>
